Rename helpWanted components
