+++ ui/pages/_manufacturerKey/_fixtureKey.vue
@@ -86,7 +86,7 @@
         </ul>
       </app-labeled-value>
 
-      <app-fixture-help-wanted
+      <app-help-wanted-message
         v-if="fixture.isHelpWanted"
         type="fixture"
         :context="fixture"
@@ -184,7 +184,7 @@
       </div>
     </section>
 
-    <app-fixture-help-wanted-dialog v-model="helpWantedContext" />
+    <app-help-wanted-dialog v-model="helpWantedContext" />
   </div>
 </template>
 
@@ -288,11 +288,11 @@ import categoryBadge from '~/components/category-badge.vue';
 import conditionalDetailsVue from '~/components/conditional-details.vue';
 import downloadButtonVue from '~/components/download-button.vue';
 import fixturePhysical from '~/components/fixture-physical.vue';
-import fixtureHelpWanted from '~/components/fixture-help-wanted.vue';
-import fixtureHelpWantedDialog from '~/components/fixture-help-wanted-dialog.vue';
 import fixtureMatrix from '~/components/fixture-matrix.vue';
 import fixtureWheel from '~/components/fixture-wheel.vue';
 import fixtureMode from '~/components/fixture-mode.vue';
+import helpWantedDialog from '~/components/help-wanted-dialog.vue';
+import helpWantedMessage from '~/components/help-wanted-message.vue';
 import labeledValueVue from '~/components/labeled-value.vue';
 
 import fixtureLinksMixin from '~/assets/scripts/fixture-links-mixin.mjs';
@@ -306,11 +306,11 @@ export default {
     'app-conditional-details': conditionalDetailsVue,
     'app-download-button': downloadButtonVue,
     'app-fixture-physical': fixturePhysical,
-    'app-fixture-help-wanted': fixtureHelpWanted,
-    'app-fixture-help-wanted-dialog': fixtureHelpWantedDialog,
     'app-fixture-matrix': fixtureMatrix,
     'app-fixture-wheel': fixtureWheel,
     'app-fixture-mode': fixtureMode,
+    'app-help-wanted-dialog': helpWantedDialog,
+    'app-help-wanted-message': helpWantedMessage,
     'app-labeled-value': labeledValueVue
   },
   mixins: [fixtureLinksMixin],
